Fix landscape styles for mobile
Problem: There was a jump with the hero text when a refrsh would
occur while in landscape mode

Fix: the cause ws from the height of the navbar being set dynamically
w/ javascript. Added a conditional to ensure there was always a value
of height given to header to offset and stop the jump.